+++ packages/sdk/src/adaptors/klima.ts
@@ -1,36 +1,58 @@
-import { ethers } from 'ethers';
-import { ComputeConfig } from '../types';
-import { KlimaCarbonRetirements } from './types';
-import KlimaCarbonRetirementsAbi from './abis/KlimaCarbonRetirements.json';
 import { formatEther } from 'ethers/lib/utils';
-import { checkComputeConfig } from '../utils';
+import { gqlFetcher } from '../utils';
 
-export default async function getKlimaData(
-  address: string,
-  computeConfig: ComputeConfig
-) {
-  checkComputeConfig('getKlimaData', computeConfig, ['polygonMainnetRpc']);
-
-  const providerMatic = new ethers.providers.JsonRpcProvider({
-    allowGzip: true,
-    url: computeConfig.polygonMainnetRpc,
-  });
-
-  const klimaCarbonRetirementsContract = new ethers.Contract(
-    '0xac298CD34559B9AcfaedeA8344a977eceff1C0Fd',
-    KlimaCarbonRetirementsAbi,
-    providerMatic
-  ) as KlimaCarbonRetirements;
-
-  const retirementTotals = await klimaCarbonRetirementsContract.retirements(
-    address
-  );
-
-  return {
-    totalRetirements: parseInt(retirementTotals.totalRetirements.toString()),
-    totalCarbonRetired: parseFloat(
-      formatEther(retirementTotals.totalCarbonRetired)
-    ),
-    totalClaimed: parseInt(retirementTotals.totalRetirements.toString()),
+interface KlimaResp {
+  data: {
+    user: null | {
+      id: string;
+      totalCarbonRetired: string;
+      totalRetirements: string;
+      retirements: {
+        id: string;
+        tokenLabel: string;
+        carbonToken: string;
+        beneficiaryAddress: string;
+        beneficiaryString: string;
+        retiredAmount: string;
+        retirementMessage: string;
+      };
+    };
   };
+}
+
+export default async function getKlimaData(address: string) {
+  const response = (await gqlFetcher(
+    'https://api.thegraph.com/subgraphs/name/anudit/klima-retirements',
+    `{
+      user(id: "${address.toLowerCase()}") {
+        id
+        totalCarbonRetired
+        totalRetirements
+        retirements {
+          id
+          tokenLabel
+          carbonToken
+          beneficiaryAddress
+          beneficiaryString
+          retiredAmount
+          retirementMessage
+        }
+      }
+    }
+    `
+  )) as KlimaResp;
+
+  if (response.data.user === null) {
+    return false;
+  } else {
+    return {
+      totalRetirements: parseInt(
+        response.data.user.totalRetirements.toString()
+      ),
+      totalCarbonRetired: parseFloat(
+        formatEther(response.data.user.totalCarbonRetired)
+      ),
+      retirements: response.data.user.retirements,
+    };
+  }
 }