+++ scripts/valistRelease.js
@@ -75,7 +75,7 @@ async function publishPackage(packageName, version) {
 
 const accountName = "theconvospace"
 let projectName = "sdk"
-let version = "0.5.3"
+let version = "0.5.4"
 
 publishPackage(`@${accountName}/${projectName}`, version).then(() => {
     process.exit(0);